Cruise and CruiseApplication states according to schema

+++ ResearchCruiseApp-API/ResearchCruiseApp-API/Application/UseCases/Cruises/DeleteCruise/DeleteCruiseHandler.cs
@@ -2,6 +2,7 @@ using MediatR;
 using ResearchCruiseApp_API.Application.Common.Models.ServiceResult;
 using ResearchCruiseApp_API.Application.ExternalServices.Persistence;
 using ResearchCruiseApp_API.Application.ExternalServices.Persistence.Repositories;
+using ResearchCruiseApp_API.Domain.Common.Enums;
 
 namespace ResearchCruiseApp_API.Application.UseCases.Cruises.DeleteCruise;
 
@@ -17,6 +18,18 @@ public class DeleteCruiseHandler(
         if (cruise is null)
             return Error.NotFound();
 
+        if (cruise.Status != CruiseStatus.New && cruise.Status != CruiseStatus.Confirmed)
+            return Error.BadRequest("Nie można już usunąć rejsu");
+        
+        if (cruise.Status == CruiseStatus.Confirmed)
+        {
+            foreach (var cruiseApplication in cruise.CruiseApplications)
+            {
+                cruiseApplication.Status = CruiseApplicationStatus.Accepted;
+            }
+        }
+      
+        
         cruisesRepository.Delete(cruise);
         await unitOfWork.Complete(cancellationToken);
